% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asBipartite.R
\name{asBipartite}
\alias{asBipartite}
\title{asBipartite()}
\usage{
asBipartite(net.igr = net.igr, part.by = NA, summary.table = TRUE,
  vertex.colors = NA)
}
\arguments{
\item{net.igr}{An igraph network.}

\item{part.by}{A metadata category (i.e. gender) by which to split the network into a strictly bipartite graph.}

\item{summary.table}{Default = TRUE. Returns a data frame, summarizing edges across the bipartite graph.}

\item{vertex.colors}{Default = NA. A length-two vector of colors. Default == c("Gold", "Tomato").}
}
\description{
asBipartite() splits an edgelist into a strictly bipartite graph around
 a specified metadata category
 requires novel.net$edge.df and net.igr from Network()
  Steps:
  1) divide character uniqnames around metadata category
     using net.igr
  2) create sub-edgelist of only bipartite edges from edge.df
     this step will find edges with exactly one element from
     one of the metadata categories
  3) outputs bipartite igraph
}
\keyword{Bipartite}
\keyword{Graph}
\keyword{Utilities}
