% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReplace.R
\name{autoReplace}
\alias{autoReplace}
\title{Auto Replace}
\usage{
autoReplace(filename, pattern, replacemt = "", local = TRUE)
}
\arguments{
\item{filename}{File name as character string, i.e. "Crusoe".}

\item{pattern}{Regular expression of text to be replaced.}

\item{replacemt}{Regular expression replacement.  Accepts all correctly formatted regular expressions.}

\item{local}{Default = TRUE. If FALSE, searches for file in google drive. If TRUE, seeks file in folder filename/.}
}
\description{
This function searches a NovNet .txt file for a correctly formatted regular expression
and substitutes a replacement.
}
\details{
It is useful for deleting page numbers, formatting chapter headings,
cleaning embedded notes, and so on.
}
\keyword{Preparation}
\keyword{Text}
