% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardf_to_edgelist.R
\name{charDf_to_edgelist}
\alias{charDf_to_edgelist}
\title{charDF_to_edgelist()}
\usage{
charDf_to_edgelist(data = novel.dat, context = 15, multiplex = FALSE,
  token.threshold = 1, metadata.off = NA, sample = FALSE,
  noisy = TRUE)
}
\arguments{
\item{data}{A data object of the format returned by txt_to_Df().  Default = novel.dat.}

\item{context}{Counting number.  How many words from each name should the function seek others?  Default = 15.}

\item{multiplex}{Logical vector of length one.  Return a multiplex graph?  Default = FALSE.}

\item{token.threshold}{Length-one numerical vector.  Removes characters with token count < x from graph.  Default = 1.}

\item{metadata.off}{Default = NA. Expects a length-two vector of c("meta", "VALUE"), or matrix of vectors, to cull characters by metadata category.  Useful for removing non-characters, generating subgraphs by gender, etc...}

\item{sample}{Default = FALSE.  A logical vector indicating whether the .txt should be randomized (i.e. words scrambled).  Useful for checking whether modularity results are statistically meaningful.}

\item{noisy}{Defaults to TRUE.}
}
\description{
Along with txt_to_DF(), this is one of the crucial NovNet utilities.
This function accepts objects from Txt_to_df()
It requires
  a) char.data.df
  b) text.uniq.v
And returns:
  1) an edgelist as a dataframe
  2) an igraph object
}
\keyword{NovNet}
\keyword{Utilities}
