% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_to_third.R
\name{complex_to_third}
\alias{complex_to_third}
\title{complex_to_third()}
\usage{
complex_to_third(input.text, narrator.breaks)
}
\arguments{
\item{input.text}{A character vector of a text, with punctuation left in.  Each element may be a line, sentence, paragraph, etc...}

\item{narrator.breaks}{A data frame with three columns: start.chap, end.chap, egoname.  It translates 1st-person pronouns in each section (start.chap to end.chap) to the associated egoname.  Each egoname must be specified in the associated Char.csv file.}
}
\description{
This function is used by txt_to_df() to transform first-person pronouns in
texts with complex narrative voices.
}
\keyword{NovNet}
\keyword{Utilities}
