% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createColorPalette.R
\name{createColorPalette}
\alias{createColorPalette}
\title{createColorPalette()}
\usage{
createColorPalette(edgelist = net.igr, edge.color.index = "weight",
  vertex.color.index = "tokenCount", color.palette = c("gold",
  "tomato"), alpha.range = c(0.5, 1), vertex.edge = "edge")
}
\arguments{
\item{edgelist}{Default = net.igr; an igraph-type graph object, from charDf_to_edgelist().}

\item{edge.color.index}{Character string. Which edge value should be used to assign colors to edges? Defaults to "weight".}

\item{vertex.color.index}{Character string. Which vertex value should used to assign colors to vertices? Defaults to "tokenCount".}

\item{color.palette}{Length-two character vector.  Default = c("gold", "tomato").}

\item{alpha.range}{Length-two numeric vector 0 <= x <= 1. Default = c(.5, 1).  Sets alpha (transparency) for edges.}

\item{vertex.edge}{Character string "vertex" or "edge".  Return values for edges or vertices?}
}
\description{
This function generates edge and vertex colors.
}
\keyword{NovNet}
\keyword{Utilities}
