% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docSetCheck.R
\name{docSetCheck}
\alias{docSetCheck}
\title{docSetCheck()}
\usage{
docSetCheck(filename, return.result = FALSE, local = FALSE,
  write.summary = TRUE, write.txtDisambig = TRUE,
  update.master = TRUE)
}
\arguments{
\item{filename}{Character string, i.e. "Crusoe"}

\item{return.result}{Logical.  If TRUE, returns a list summary of results.}

\item{local}{If FALSE (default), downloads from Google Drive and saves to folder.}

\item{write.summary}{Logical.  If TRUE, writes summary in filename/filename-docSetCheck.txt.}

\item{write.txtDisambig}{Logical. Passed to txtDisambig(). If TRUE, writes detailed report at filename/filename-txtDisambig.txt.}

\item{update.master}{Logical.  If TRUE looks for Results/docSetCheck.csv, and updates the relevant line.}
}
\description{
This function accepts a filename and:
a) Checks for a complete document set.
b) Checks for orphaned names and duplicates.
c) Checks for a few other problems (like badly formed metadata categories)
d) Returns a summary as a list
}
\keyword{NovNet}
\keyword{Utilities}
