% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docSetComplete.R
\name{docSetComplete}
\alias{docSetComplete}
\title{docSetComplete()
pulls all files with .txt and Char.csv
runs docSetCheck() for all files
writes docSetCheck.csv}
\usage{
docSetComplete(local = TRUE)
}
\arguments{
\item{local}{If FALSE (default), downloads from Google Drive and saves to folder.}
}
\description{
docSetComplete()
pulls all files with .txt and Char.csv
runs docSetCheck() for all files
writes docSetCheck.csv
}
\keyword{NovNet}
\keyword{Utilities}
