% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistle_to_third.R
\name{epistle_to_third}
\alias{epistle_to_third}
\title{epistle_to_third()}
\usage{
epistle_to_third(input.text, author.name, addressee.name)
}
\arguments{
\item{input.text}{A character vector of a single epistle, with punctuation left in.  Each element may be a line, sentence, paragraph, etc...}

\item{author.name}{A character string of the alternate name token to signal the first-person author of an epistle.  This function transforms non-dialogic 1st-person pronouns into this variable's value.  It must therefore match a unique alternate name on the associated Char.csv file.}

\item{addressee.name}{A character string of the alternate name token to signal the second-person recipient of an epistle.  This function transforms 2nd-person pronouns into this variable's value.  It must therefore match a unique alternate name on the associated Char.csv file.}
}
\description{
This function is used by txt_to_df() to transform first-person pronouns in
texts with first-person narrative voice.
}
\keyword{NovNet}
\keyword{Utilities}
