% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistolary.R
\name{epistolary}
\alias{epistolary}
\title{epistolary()
This function is used by Txt_to_df() to transform epistolary authors and addressees into chracter names.
1) Splits input text.v into epistles
2) Applies first_to_third() selectively, to transform first and second
   person address in epistles into sender and addressee uniqnames}
\usage{
epistolary(input.text = main.text.v)
}
\arguments{
\item{input.text}{A text transformed into a character vector; each element is a line of text. Txt_to_df() returns a correctly formatted vector as $main.text.v.}
}
\description{
REQUIRES: a) *Epistle.csv, a list of letters with sender and addressee
             Filename must match .txt file
          b) Letters headers in .txt on their own line, at the start of the line, in one of two formats:
             EPISTLE [number]
             LETTER [number]
}
\keyword{NovNet}
\keyword{Utilities}
