% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_to_third.R
\name{first_to_third}
\alias{first_to_third}
\title{first_to_third()}
\usage{
first_to_third(input.text, egoname)
}
\arguments{
\item{input.text}{A character vector of a text, with punctuation left in.  Each element may be a line, sentence, paragraph, etc...}

\item{egoname}{A character string of the alternate name token to signal the first-person narrator.  This function transforms non-dialogic 1st-person pronouns into this variable's value (i.e. "EgoName", etc...).  It must therefore match a unique alternate name on the associated Char.csv file.}
}
\description{
This function is used by txt_to_df() to transform first-person pronouns in
texts with first-person narrative voice.
}
\keyword{NovNet}
\keyword{Utilities}
