% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linesToParas.R
\name{linesToParas}
\alias{linesToParas}
\title{Lines to Paragraphs}
\usage{
linesToParas(filename, write.out = TRUE, return.text = FALSE,
  scan.file = TRUE, local = TRUE)
}
\arguments{
\item{filename}{File name as character string, i.e. "Crusoe", or (if scan.file = FALSE) a character vector of a text where each string is a line of text.}

\item{write.out}{Logical.  If TRUE (default), writes to file filename/filename-linesToParas.txt.}

\item{return.text}{Logical.  If TRUE (default = FALSE), returns the transformed text as a vector.}

\item{scan.file}{Logical.  If TRUE (default), scans in filename from filename/filename.txt, or (if local = FALSE) Google Drive filename.txt.}

\item{local}{If FALSE (default), downloads from Google Drive and saves to folder filename/.}
}
\description{
This function accepts a .txt file with Project Gutenberg format
and removes inner-paragraph line breaks.
}
\keyword{Preparation}
\keyword{Text}
