% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{Network}
\alias{Network}
\title{Network()
This function bundles together steps following charDf_to_edgelist().}
\usage{
Network(novel.net = novel.net, layout = "fr", remove.isolates = TRUE,
  min.degree = 0)
}
\arguments{
\item{novel.net}{List including igraph network, produced by charDf_to_edgelist().  Default is novel.net.}

\item{layout}{FALSE, or Character, passed to igraph layout function layout_with_XXX(). Accepts fr, "gem", kk, tree, circle, sphere, lgl. If FALSE, no layout is attached.}

\item{remove.isolates}{Logical.  Passed to chardf_to_edgelist.  If TRUE, isolates are removed}

\item{min.degree}{Numeric.  Vertices of lower degree are ignored.}
}
\description{
It requires the list returned by charDf_to_edgelist (novel.net).
It simplifies, removes isolates, and attaches a layout as a global variable
}
\keyword{NovNet}
\keyword{Utilities}
