% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFabric.R
\name{nnFabric}
\alias{nnFabric}
\title{nnFabric()}
\usage{
nnFabric(filename, context = 10, set.voice = NA, chunk.length = 5000,
  include.all = TRUE, brewer.palette = "YlGnBu",
  mirror.colors = FALSE, bg.col = "white", thread.edge = "black",
  thread.width = 1, knot.size = 2, local = FALSE, characters = NA,
  token.thresh = 1)
}
\arguments{
\item{argument}{Description ends with .}
}
\description{
Creates a "Fabric" visualization from a novel, where:
Warp yarns (horizontal) are character appearances by chapter
Weft yarns connect appearances by name mentions within a context field
}
\keyword{NovNet}
\keyword{Utilities}
