% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NovelCompiler.R
\name{NovelCompiler}
\alias{NovelCompiler}
\title{compDegreeDist()}
\usage{
NovelCompiler(token.threshold = 1)
}
\arguments{
\item{token.threshold}{Counting number, passed to charDf_to_edgelist. Removes characters of token-count less than value.  Default = 1.}
}
\description{
This script generates .igr objects for all the novels stored locally
And enables comparative analyses
It outputs a list of lists (of lists) of the format
novels.l$filename
each $filename contains: 1) novel.igr
                         2) Txt_to_df.l
                         3) charDf_to_edgelist.l
}
\keyword{NovNet}
\keyword{Utilities}
