% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBrackets.R
\name{removeBrackets}
\alias{removeBrackets}
\title{removeBrackets()}
\usage{
removeBrackets(filename, local = TRUE)
}
\arguments{
\item{filename}{File name as character string, i.e. "Crusoe".}

\item{local}{Default = TRUE. If FALSE, searches for file in google drive. If TRUE, seeks file in folder filename/.}
}
\description{
2) removeBrackets() Remove bracketed names ----
this is a companion function to txtDisambig().  It:
a) searches for stray bracketed names.
b) strips out stray brackets.
c) locates the errata section
   i) removes the errata
   iii) saves as .txt file
}
\keyword{Disambiguation}
