% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romanToArabic.R
\name{romanToArabic}
\alias{romanToArabic}
\title{Roman Numeral Conversion}
\usage{
romanToArabic(filename)
}
\arguments{
\item{filename}{File name as character string, i.e. "Crusoe".}
}
\description{
This function converts roman numeral chapter numbers to arabic, from Project Gutenberg files.
It expects chapter headings in this format: CHAPTER I, CHAPTER II, etc...
}
\keyword{Disambiguation}
