% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagNames.R
\name{tagNames}
\alias{tagNames}
\title{tagNames()
tagNames() produces a .txt file with all found alternate names
  tagged with NovNet-style tags:
  < Alternate Name u: uniqname \>}
\usage{
tagNames(filename, local = FALSE, include.all = TRUE)
}
\arguments{
\item{filename}{Character.  Filename as single character string, i.e. "Crusoe".}

\item{local}{If FALSE (default), downloads from Google Drive and saves to folder.}

\item{include.all}{Logical vector or numerical range. If TRUE (default), processes whole text.  If range (i.e. 1:18), processes only those chapters.  If FALSE, will offer a prompt to determine a range.}
}
\description{
tagNames()
tagNames() produces a .txt file with all found alternate names
  tagged with NovNet-style tags:
  < Alternate Name u: uniqname \>
}
\keyword{NovNet}
\keyword{PhaseII}
\keyword{Utilities}
